factor out the Flusher in its own module, while reusing code
